+++ src/main/java/com/gmail/val59000mc/listeners/BlockListener.java
@@ -8,23 +8,25 @@ import com.gmail.val59000mc.game.GameManager;
 import com.gmail.val59000mc.languages.Lang;
 import com.gmail.val59000mc.players.PlayerManager;
 import com.gmail.val59000mc.players.UhcPlayer;
-import com.gmail.val59000mc.utils.RandomUtils;
 import com.gmail.val59000mc.utils.UniversalMaterial;
 import org.bukkit.Bukkit;
 import org.bukkit.GameMode;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
+import org.bukkit.enchantments.Enchantment;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockBreakEvent;
+import org.bukkit.event.block.BlockEvent;
 import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.event.block.LeavesDecayEvent;
 import org.bukkit.inventory.ItemStack;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.ThreadLocalRandom;
 
 public class BlockListener implements Listener{
 
@@ -51,11 +53,6 @@ public class BlockListener implements Listener{
 	public void onBlockPlace(BlockPlaceEvent event){
 		handleFrozenPlayers(event);
 		handleMaxBuildingHeight(event);
-	}
-
-	@EventHandler
-	public void onLeavesDecay(LeavesDecayEvent event){
-		handleAppleDrops(event);
 	}
 
 	private void handleMaxBuildingHeight(BlockPlaceEvent e){
@@ -91,24 +88,34 @@ public class BlockListener implements Listener{
 		}
 	}
 
-	// High priority so that custom block drop scenarios can override it
-	@EventHandler(priority = EventPriority.HIGH)
-	public void handleShearedLeaves(BlockBreakEvent e) {
+	// Low priority so that extra apples can still drop in custom block drop scenarios
+	@EventHandler(priority = EventPriority.LOW)
+	public void onPlayerLeavesBreak(BlockBreakEvent e) {
 		if (e.isCancelled()) {
 			return;
 		}
 
-		if (!configuration.get(MainConfig.APPLE_DROPS_FROM_SHEARING)){
+		if (!UniversalMaterial.isLeaves(e.getBlock().getType())) {
 			return;
 		}
 
-		if (!UniversalMaterial.isLeaves(e.getBlock().getType())){
+		final ItemStack itemInMainHand = e.getPlayer().getItemInHand();
+		if (itemInMainHand.getType() == Material.SHEARS && !configuration.get(MainConfig.APPLE_DROPS_FROM_SHEARING)) {
+			return;
+		}
+		if (itemInMainHand.getItemMeta().hasEnchant(Enchantment.SILK_TOUCH)) {
 			return;
 		}
 
-		if (e.getPlayer().getItemInHand().getType() == Material.SHEARS){
-			Bukkit.getPluginManager().callEvent(new LeavesDecayEvent(e.getBlock()));
+		handleAppleDrops(e);
+	}
+
+	@EventHandler
+	public void onLeavesDecay(LeavesDecayEvent e) {
+		if (e.isCancelled()) {
+			return;
 		}
+		handleAppleDrops(e);
 	}
 
 	private void handleFrozenPlayers(BlockBreakEvent e){
@@ -133,43 +140,26 @@ public class BlockListener implements Listener{
 		}
 	}
 
-	private void handleAppleDrops(LeavesDecayEvent e){
-		if (e.isCancelled()) {
+	private void handleAppleDrops(BlockEvent e) {
+		final Block block = e.getBlock();
+		final boolean shouldDropExtraApple = UniversalMaterial.isAppleLeaves(block) || configuration.get(MainConfig.APPLE_DROPS_FROM_ALL_TREES);
+
+		if (!shouldDropExtraApple) {
 			return;
 		}
 
-		Block block = e.getBlock();
-		Material type = block.getType();
-		boolean isOak;
-
-		if (configuration.get(MainConfig.APPLE_DROPS_FROM_ALL_TREES)){
-			if (type != UniversalMaterial.OAK_LEAVES.getType()) {
-				e.getBlock().setType(UniversalMaterial.OAK_LEAVES.getType());
-			}
-			isOak = true;
-		}else {
-			isOak = type == UniversalMaterial.OAK_LEAVES.getType() || type == UniversalMaterial.DARK_OAK_LEAVES.getType();
+		final double dropPercentage = configuration.get(MainConfig.APPLE_DROP_PERCENTAGE);
+		if (dropPercentage <= 0) {
+			return;
 		}
 
-		if (!isOak){
-			return; // Will never drop apples so drops don't need to increase
+		final double random = ThreadLocalRandom.current().nextDouble(100);
+		if (random < dropPercentage) {
+			// Add apple to drops
+			Bukkit.getScheduler().runTask(UhcCore.getPlugin(),
+				() -> block.getWorld().dropItem(block.getLocation().add(.5, .5, .5),
+			new ItemStack(Material.APPLE)));
 		}
-
-		double percentage = configuration.get(MainConfig.APPLE_DROP_PERCENTAGE)-0.5;
-
-		if (percentage <= 0){
-			return; // No added drops
-		}
-
-		// Number 0-100
-		double random = RandomUtils.randomInteger(0, 200)/2D;
-
-		if (random > percentage){
-			return; // Number above percentage so no extra apples.
-		}
-
-		// Add apple to drops
-		Bukkit.getScheduler().runTask(UhcCore.getPlugin(), () -> block.getWorld().dropItem(block.getLocation().add(.5, .5, .5), new ItemStack(Material.APPLE)));
 	}
 
 }