+++ CHANGELOG.md
@@ -4,14 +4,22 @@
 
 ### New features
 
+- [#190][GL-190] - Clarify apple-drops option behavior
 - [#182][GL-182] - Option to change scoreboard render type for the hearts-on-tab option
 
 ### Bug fixes
 
+- [#191][GL-191] - Extra apples from the apple-drops option are only dropped when leaves decay and not when a player breaks leaves
+- [#192][GL-192] - All types of decaying leaves drop oak saplings when the apple-drops.all-trees option is enabled
+- [#193][GL-193] - The apple-drops.all-trees option can only be disabled on Minecraft 1.13+
 - [#184][GL-184] - Various bugs and compatibility issues between scenarios with custom block drops
 - [#180][GL-180] - Player death message is not client-translatable on Paper servers for Minecraft 1.16.5+
 - [#177][GL-177] - Iteminfo command fails when using it on a potion that has no base effect on Minecraft 1.20.5+
 
+[GL-193]: https://gitlab.com/uhccore/uhccore/-/issues/193
+[GL-192]: https://gitlab.com/uhccore/uhccore/-/issues/192
+[GL-191]: https://gitlab.com/uhccore/uhccore/-/issues/191
+[GL-190]: https://gitlab.com/uhccore/uhccore/-/issues/190
 [GL-184]: https://gitlab.com/uhccore/uhccore/-/issues/184
 [GL-182]: https://gitlab.com/uhccore/uhccore/-/issues/182
 [GL-180]: https://gitlab.com/uhccore/uhccore/-/issues/180